# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  common.ResponseError:
    properties:
      error:
        type: string
    type: object
...
  common.ResponseNormal:
    properties:
      data: {}
      email:
        type: string
      message:
        type: string
      paging: {}
      result:
        type: boolean
    type: object
  dtos.CreateActivity:
    properties:
      description:
        type: string
      duration:
        type: integer
      point_reward:
        type: integer
      repeat_days:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        enum:
        - Activity
        - MiniGame
        - Challenge
        type: string
    required:
    - duration
    - point_reward
    - title
    - type
    type: object
...
  dtos.CreateLevel:
    properties:
      activities:
        items:
          $ref: '#/definitions/dtos.CreateActivity'
        type: array
      description:
        type: string
      name:
        type: string
      point_require:
        type: integer
    required:
    - name
    - point_require
    type: object
...
  dtos.CreateProgramRequest:
    properties:
      desciption:
        type: string
      disease_ids:
        items:
          type: integer
        type: array
      duration:
        type: integer
      goal_ids:
        items:
          type: integer
        type: array
      levels:
        items:
          $ref: '#/definitions/dtos.CreateLevel'
        type: array
      title:
        type: string
    required:
    - duration
    - title
    type: object
  dtos.UpdateProgramRequest:
    properties:
      desciption:
        type: string
      disease_ids:
        items:
          type: integer
        type: array
      duration:
        type: integer
      goal_ids:
        items:
          type: integer
        type: array
      title:
        type: string
    required:
    - duration
    - title
    type: object
...
  models.Program:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      duration:
        type: integer
      is_active:
        type: boolean
      program_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    required:
    - duration
    - title
    type: object
host: 127.0.0.1:9000
info:
  contact: {}
  description: List APIs of Healthy Management Service
  termsOfService: http://swagger.io/terms/
  title: Healthy Service API Document
  version: "1.0"
paths:
...
  /expert/programs:
    get:
      consumes:
      - application/json
      description: Get programs by expert ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Programs retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseNormal'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Program'
                  type: array
              type: object
        "400":
          description: Invalid expert ID format
          schema:
            $ref: '#/definitions/common.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get programs by expert ID
      tags:
      - Program
    post:
      consumes:
      - application/json
      description: Create a new program
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Program request data
        in: body
        name: program
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateProgramRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Program created successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseNormal'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/common.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create a new program
      tags:
      - Program
  /expert/programs/{program_id}:
    delete:
      consumes:
      - application/json
      description: Deletes a program if it has no participants, or deactivates it
        (sets is_active = false)
      parameters:
      - description: Program ID (UUID)
        format: uuid
        in: path
        name: program_id
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Programs retrieved successfully
          schema:
            $ref: '#/definitions/common.ResponseNormal'
        "400":
          description: invalid program ID or confirmation required
          schema:
            $ref: '#/definitions/common.ResponseError'
        "401":
          description: token expired or invalid
          schema:
            $ref: '#/definitions/common.ResponseError'
        "403":
          description: 'unauthorized: not program owner'
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: failed to delete program
          schema:
            $ref: '#/definitions/common.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Delete or deactivate a program by ID
      tags:
      - Program
    put:
      consumes:
      - application/json
      description: Updates program details (title, description, duration, disease
        IDs, goal IDs) for a program owned by the expert. Requires valid JWT token.
      parameters:
      - description: Program ID (UUID)
        format: uuid
        in: path
        name: program_id
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Program details to update
        in: body
        name: program
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateProgramRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Program updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseNormal'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: invalid program ID or invalid program data
          schema:
            $ref: '#/definitions/common.ResponseError'
        "401":
          description: token expired or invalid
          schema:
            $ref: '#/definitions/common.ResponseError'
        "403":
          description: 'unauthorized: not program owner'
          schema:
            $ref: '#/definitions/common.ResponseError'
        "404":
          description: expert not found or program not found
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: failed to update program
          schema:
            $ref: '#/definitions/common.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Update a program by ID
      tags:
      - Program
  /profile:
    post:
      consumes:
      - multipart/form-data
      description: Create user profile with file image and json profile
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Profile image file (max 10MB)
        in: formData
        name: image
        type: file
      - description: Json body for profile
        in: formData
        name: metadata
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Profile created successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseNormal'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: invalid request form-data
          schema:
            $ref: '#/definitions/common.ResponseError'
        "401":
          description: invalid token
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ResponseError'
      summary: Create a new profile
      tags:
      - Profile
  /profile/{id}:
    put:
      consumes:
      - multipart/form-data
      description: Update user profile with file image and json profile
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Profile image file (max 10MB)
        in: formData
        name: image
        type: file
      - description: Json body for profile
        in: formData
        name: metadata
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseNormal'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: invalid request form-data
          schema:
            $ref: '#/definitions/common.ResponseError'
        "401":
          description: invalid token
          schema:
            $ref: '#/definitions/common.ResponseError'
        "404":
          description: Profile not found
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ResponseError'
      summary: Update user profile
      tags:
      - Profile
schemes:
- http
- https
swagger: "2.0"